feat: deep clean from tiktok 👍
